requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.19.0
anyio==4.2.0
attrs==23.2.0
backoff==2.2.1
botocore==1.34.38
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
gql==3.5.0
graphql-core==3.2.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iso8601==2.1.0
jmespath==1.0.1
multidict==6.0.5
numpy==1.26.4
psutil==5.9.8
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.0
twitchio==2.8.2
typing_extensions==4.9.0
urllib3==2.0.7
websockets==11.0.3
yarl==1.9.4
